Allow the Starting Image to be set by drag and drop (#335)
* img2img image well drag and drop

- duplicate functionality from ControlNetView image well
  - drag and drop images
  - click image well to open file picker

* Refactor ImageView into portable ImageWellView

* Move ImageWellView to its own file

* Use ImageWellView in ControlNetView

* Clean up unnecessary VStack

* Dragging an image to a loaded well replaces the old image with the new
