+++ Mochi Diffusion/Views/SidebarControls/ControlNetView.swift
@@ -16,57 +16,15 @@ struct ControlNetView: View {
             .sidebarLabelFormat()
 
         HStack(alignment: .top) {
-            if let image = controller.currentControlNets.first?.image {
-                Image(image, scale: 1, label: Text(verbatim: ""))
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .padding(3)
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 2)
-                            .stroke(Color(nsColor: .separatorColor), lineWidth: 1)
-                    )
-                    .frame(height: 90)
-                    .accessibilityAddTraits(.isButton)
-            } else {
-                Button {
-                    Task { await ImageController.shared.selectControlNetImage(at: 0) }
-                } label: {
-                    Image(systemName: "photo")
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .foregroundColor(Color(nsColor: .separatorColor))
-                        .padding(30)
-                        .overlay(
-                            RoundedRectangle(cornerRadius: 2)
-                                .stroke(Color(nsColor: .separatorColor), lineWidth: 1)
-                        )
-                        .background(.background.opacity(0.01))
-                        .frame(height: 90)
-                        .onDrop(of: [.fileURL], isTargeted: nil) { providers in
-                            _ = providers.first?.loadDataRepresentation(for: .fileURL) { data, _ in
-                                guard let data, let urlString = String(data: data, encoding: .utf8), let url = URL(string: urlString) else {
-                                    return
-                                }
-
-                                guard let cgImageSource = CGImageSourceCreateWithURL(url as CFURL, nil) else {
-                                    return
-                                }
-
-                                let imageIndex = CGImageSourceGetPrimaryImageIndex(cgImageSource)
-
-                                guard let cgImage = CGImageSourceCreateImageAtIndex(cgImageSource, imageIndex, nil) else {
-                                    return
-                                }
-
-                                Task { await ImageController.shared.setControlNet(image: cgImage) }
-                            }
-
-                            return true
-                        }
+            ImageWellView(image: controller.currentControlNets.first?.image) { image in
+                if let image {
+                    await ImageController.shared.setControlNet(image: image)
+                } else {
+                    await ImageController.shared.unsetControlNet()
                 }
-                .buttonStyle(.plain)
-                .disabled(controller.controlNet.isEmpty)
             }
+            .disabled(controller.controlNet.isEmpty)
+            .frame(height: 90)
 
             Spacer()
